requirements.txt
pyaudio
SpeechRecognition
google-genai
python-dotenv
pyttsx3
pywin32
keyboard
cryptography
pymysql
redis
rich